fixes #177 -- replace markdown and wysiwyg with new editor
