+++ app/assets/javascripts/builder/application.js
@@ -3,8 +3,8 @@
 //= require jquery-ui/sortable
 //= require jquery-fileupload/basic
 //= require jquery-fileupload/vendor/tmpl
-//= require wysihtml5
-//= require builder/wysihtml5_custom
+//= require editor
+//= require marked
 //= require pickadate/picker
 //= require pickadate/picker.date
 //= require pickadate/picker.time
@@ -13,7 +13,5 @@
 //= require backbone
 //= require ./app
 //= require_tree ./templates
-//= require_tree ./models
-//= require_tree ./collections
 //= require_tree ./views
 //= require_tree ./routers
