--- app/assets/stylesheets/builder/partials/markdown_editor.scss
@@ -1,121 +1,0 @@
-.ajax-page.markdown-editor {
-
-  .ajax-page-content {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    margin: 0;
-    padding: 50px 0 0 0;
-    font-size: 0;
-    overflow: hidden;
-
-    .half {
-      display: inline-block;
-      width: 50%;
-      height: 100%;
-      // padding-top: 68px;
-      @include box-sizing(border-box);
-      overflow: scroll;
-
-      &.left {
-        overflow: hidden;
-
-        form {
-          position: relative;
-          height: 100%;
-          &.simple_form.new_image {
-            display: none;
-          }
-        }
-        .toolbar {
-          position: fixed;
-          top: -50%;
-          width: 50%;
-          max-height: 50%;
-          overflow-y: scroll;
-          background-color: $color-light-darker;
-          z-index: 99;
-          @include transition(all 0.35s ease);
-          &.active {
-            // top: 68px;
-            top: 50px;
-          }
-          &.form {
-            margin: 0;
-            padding: 0 5%;
-            label {
-              display: block;
-              margin: 10px 0 0 0;
-              font-size: 14px;
-              text-transform: uppercase;
-              color: $color-1;
-              font-weight: 400;
-            }
-            input {
-              margin: 0 0 10px 0;
-            }
-          }
-          .image-library {
-            margin: 10px auto;
-            text-align: center;
-            font-size: 0;
-            img {
-              display: inline-block;
-              width: 90px;
-              margin: 5px;
-              border: 3px solid transparent;
-              &:hover {
-                cursor: pointer;
-                border-color: $color-2;
-              }
-            }
-          }
-          ul {
-            margin: 0;
-            padding: 0;
-            li {
-              display: inline-block;
-              margin: 10px;
-              a {
-                display: block;
-                padding: 5px;
-                border: 1px solid $color-light-darkest;
-                border-radius: 3px;
-                line-height: 14px;
-                i {
-                  color: $color-light-darkest;
-                  font-size: 14px;
-                }
-              }
-            }
-          }
-        }
-        textarea {
-          position: absolute;
-          width: 100%;
-          bottom: 0;
-          height: 100%;
-          margin: 0;
-          // padding: 66px 20px 20px 20px;
-          padding: 20px;
-          border-color: $color-light-darker;
-          background-color: $color-light-dark;
-          font-family: "Droid Sans Mono", monospace, inherit;
-          font-size: 12px;
-          line-height: 18px;
-          color: $color-dark;
-          &:focus {
-            border-color: $color-light-darkest;
-          }
-        }
-      }
-
-      &.right {
-        // padding: 68px 20px 20px 20px;
-        padding: 20px;
-      }
-    }
-  }
-}
