--- app/assets/stylesheets/builder/wysihtml5.css
@@ -1,7 +1,0 @@
-a {
-  color: #00ade2;
-  text-decoration: none;
-}
-a:hover {
-  text-decoration: underline;
-}
