+++ app/views/layouts/builder.html.erb
@@ -15,11 +15,6 @@
 
     <%= csrf_meta_tags %>
 
-    <script type="text/javascript">
-      var WYSIWYG_BASE_STYLES = "<%= stylesheet_path('wysiwyg') %>";
-      var WYSIWYG_CUSTOM_STYLES = "<%= stylesheet_path('builder/wysihtml5') %>";
-    </script>
-
     <% if Rails.env.production? && SapwoodSetting.google.analytics_id.present? %>
       <script>
         (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
