+++ config/routes.rb
@@ -52,8 +52,6 @@ Rails.application.routes.draw do
         end
         get 'move' => 'pages#move', :as => :move
         get 'settings/:slug' => 'pages#edit', :as => :settings
-        get 'edit/:editor' => 'pages/editor#edit', :as => :editor
-        patch 'edit/:editor' => 'pages/editor#parse', :as => :parser
         get 'help' => 'pages#help', :as => :help
         post 'publish' => 'pages#publish', :as => :publish
         post 'unpublish' => 'pages#unpublish', :as => :unpublish
